+++ src/maicos/core/sphere.py
@@ -32,8 +32,7 @@ class SphereBase(AnalysisBase):
     ----------
     ${ATOMGROUPS_PARAMETER}
     ${SPHERE_CLASS_PARAMETERS}
-    kwargs : dict
-        Parameters parsed to `AnalysisBase`.
+    ${WRAP_COMPOUND_PARAMETER}
 
     Attributes
     ----------
@@ -61,12 +60,25 @@ class SphereBase(AnalysisBase):
     def __init__(
         self,
         atomgroups: Union[mda.AtomGroup, List[mda.AtomGroup]],
+        multi_group: bool,
+        unwrap: bool,
+        refgroup: Optional[mda.AtomGroup],
+        jitter: float,
+        concfreq: int,
         rmin: float,
         rmax: Union[None, float],
         bin_width: float,
-        **kwargs,
+        wrap_compound: str,
     ):
-        super().__init__(atomgroups, **kwargs)
+        super().__init__(
+            atomgroups=atomgroups,
+            multi_group=multi_group,
+            unwrap=unwrap,
+            refgroup=refgroup,
+            jitter=jitter,
+            concfreq=concfreq,
+            wrap_compound=wrap_compound,
+        )
 
         self.rmin = rmin
         self._rmax = rmax
@@ -137,8 +149,8 @@ class ProfileSphereBase(SphereBase, ProfileBase):
 
     Parameters
     ----------
-    ${PROFILE_CLASS_PARAMETERS_PRIVATE}
     ${PROFILE_SPHERE_CLASS_PARAMETERS}
+    ${PROFILE_CLASS_PARAMETERS_PRIVATE}
 
     Attributes
     ----------
@@ -147,33 +159,45 @@ class ProfileSphereBase(SphereBase, ProfileBase):
 
     def __init__(
         self,
-        weighting_function: Callable,
-        normalization: str,
         atomgroups: Union[mda.AtomGroup, List[mda.AtomGroup]],
+        unwrap: bool,
+        refgroup: Optional[mda.AtomGroup],
+        jitter: float,
+        concfreq: int,
+        rmin: float,
+        rmax: Union[None, float],
+        bin_width: float,
         grouping: str,
         bin_method: str,
         output: str,
-        f_kwargs: Optional[Dict] = None,
-        **kwargs,
+        weighting_function: Callable,
+        weighting_function_kwargs: Optional[Dict],
+        normalization: str,
     ):
         SphereBase.__init__(
             self,
             atomgroups=atomgroups,
+            unwrap=unwrap,
+            refgroup=refgroup,
+            jitter=jitter,
             multi_group=True,
-            wrap_compound=grouping,
-            **kwargs,
+            concfreq=concfreq,
+            rmin=rmin,
+            rmax=rmax,
+            bin_width=bin_width,
+            wrap_compound=grouping,  # same as grouping to avoid broken compounds
         )
         # `AnalysisBase` performs conversions on `atomgroups`.
         # Take converted `atomgroups` and not the user provided ones.
         ProfileBase.__init__(
             self,
             atomgroups=self.atomgroups,
-            weighting_function=weighting_function,
-            normalization=normalization,
             grouping=grouping,
             bin_method=bin_method,
             output=output,
-            f_kwargs=f_kwargs,
+            weighting_function=weighting_function,
+            weighting_function_kwargs=weighting_function_kwargs,
+            normalization=normalization,
         )
 
     def _prepare(self):