+++ src/maicos/modules/densityplanar.py
@@ -60,13 +60,13 @@ class DensityPlanar(ProfilePlanarBase):
         output: str = "density.dat",
         concfreq: int = 0,
         jitter: float = 0.0,
-    ):
+    ) -> None:
         self._locals = locals()
         super().__init__(
-            weighting_function=density_weights,
-            f_kwargs={"dens": dens},
-            normalization="volume",
             atomgroups=atomgroups,
+            unwrap=unwrap,
+            jitter=jitter,
+            concfreq=concfreq,
             dim=dim,
             zmin=zmin,
             zmax=zmax,
@@ -74,9 +74,9 @@ class DensityPlanar(ProfilePlanarBase):
             refgroup=refgroup,
             sym=sym,
             grouping=grouping,
-            unwrap=unwrap,
             bin_method=bin_method,
             output=output,
-            concfreq=concfreq,
-            jitter=jitter,
+            weighting_function=density_weights,
+            weighting_function_kwargs={"dens": dens},
+            normalization="volume",
         )