+++ src/maicos/modules/dipordercylinder.py
@@ -58,7 +58,7 @@ class DiporderCylinder(ProfileCylinderBase):
         pdim: str = "r",
         order_parameter: str = "P0",
         jitter: float = 0.0,
-    ):
+    ) -> None:
         if order_parameter == "P0":
             normalization = "volume"
         else:
@@ -74,24 +74,24 @@ class DiporderCylinder(ProfileCylinderBase):
             )
 
         super().__init__(
-            weighting_function=diporder_weights,
-            f_kwargs={
-                "order_parameter": order_parameter,
-                "get_unit_vectors": get_unit_vectors,
-            },
-            normalization=normalization,
             atomgroups=atomgroups,
+            unwrap=unwrap,
+            refgroup=refgroup,
+            jitter=jitter,
+            concfreq=concfreq,
             dim=dim,
             zmin=zmin,
             zmax=zmax,
             bin_width=bin_width,
             rmin=rmin,
             rmax=rmax,
-            refgroup=refgroup,
             grouping=grouping,
-            unwrap=unwrap,
             bin_method=bin_method,
             output=output,
-            concfreq=concfreq,
-            jitter=jitter,
+            weighting_function=diporder_weights,
+            weighting_function_kwargs={
+                "order_parameter": order_parameter,
+                "get_unit_vectors": get_unit_vectors,
+            },
+            normalization=normalization,
         )