+++ src/maicos/modules/velocitycylinder.py
@@ -61,7 +61,7 @@ class VelocityCylinder(ProfileCylinderBase):
         jitter: float = 0.0,
         vdim: int = 0,
         flux: bool = False,
-    ):
+    ) -> None:
         self._locals = locals()
         if vdim not in [0, 1, 2]:
             raise ValueError("Velocity dimension can only be x=0, y=1 or z=2.")
@@ -71,10 +71,10 @@ class VelocityCylinder(ProfileCylinderBase):
             normalization = "number"
 
         super(VelocityCylinder, self).__init__(
-            weighting_function=velocity_weights,
-            f_kwargs={"vdim": vdim},
-            normalization=normalization,
             atomgroups=atomgroups,
+            unwrap=unwrap,
+            jitter=jitter,
+            concfreq=concfreq,
             dim=dim,
             zmin=zmin,
             zmax=zmax,
@@ -83,9 +83,9 @@ class VelocityCylinder(ProfileCylinderBase):
             rmax=rmax,
             refgroup=refgroup,
             grouping=grouping,
-            unwrap=unwrap,
             bin_method=bin_method,
             output=output,
-            concfreq=concfreq,
-            jitter=jitter,
+            weighting_function=velocity_weights,
+            weighting_function_kwargs={"vdim": vdim},
+            normalization=normalization,
         )